+++ example/lib/forms/order_form.dart
@@ -1,3 +1,4 @@
+import 'package:auto_form_example/api/fake_api.dart';
 import 'package:flutter_auto_form/flutter_auto_form.dart';
 
 class OrderForm extends TemplateForm {
@@ -22,5 +23,11 @@ class OrderForm extends TemplateForm {
       textBuilder: (e) => e.toString(),
       value: 1,
     ),
+    AFSearchMultipleModelsField(
+      id: 'books',
+      name: 'Books',
+      validators: [NotEmptyListValidator('Please select at least one book')],
+      search: searchBook,
+    )
   ];
 }