+++ lib/src/widgets/fields/select_field.dart
@@ -24,7 +24,7 @@ class _SelectFieldWidgetState<T extends Object>
   @override
   Widget build(BuildContext context) {
     return DropdownButtonFormField<T>(
-      validator: field.validate,
+      validator: field.validator,
       autovalidateMode: widget.fieldContext.forceErrorDisplay
           ? AutovalidateMode.always
           : AutovalidateMode.onUserInteraction,